code refactored and fixed
